Add new integration test and reorganize folder structure
