+++ app.mjs
@@ -7,10 +7,10 @@ import exphbs from 'express-handlebars'
 import { indexRoutes } from './routes/index-routes.mjs'
 import { detailsRoutes } from './routes/details-routes.mjs'
 import { helpers } from './utils/handlebar-util.mjs'
-import { sessionUserSettings } from './utils/session-middleware.index.mjs'
+import { sessionUserSettings } from './utils/session-middleware.mjs'
 import { overrideMiddleware } from './utils/method-override.mjs'
-import { notFound } from './utils/not-found-middleware.index.mjs'
-import { errorHandler } from './utils/error-handler-middleware.index.mjs'
+import { notFound } from './utils/not-found-middleware.mjs'
+import { errorHandler } from './utils/error-handler-middleware.mjs'
 
 export const app = express()
 const hbs = exphbs.create({
