+++ tests/integration-tests/details-controller.test.mjs
@@ -14,14 +14,11 @@ process.env.NODE_ENV = 'testing'
 dotenv.config({ path: `.env-testing` })
 
 // load app after env
-const app = (await import('../app.mjs')).app
+const app = (await import('../../app.mjs')).app
 
 describe('GET /', () => {
-    it('should return index page', async () => {
-        const response = await chai.request(app).get('/')
-        response.should.have.status(200)
-
-        const dom = new jsdom.JSDOM(response.text)
-        expect(dom.window.document.body.innerHTML).contain('Hello World')
+    it('should return 404 page', async () => {
+        const response = await chai.request(app).get('/details/wrong_id')
+        response.should.have.status(404)
     })
 })
