+++ utils/not-found-middleware.mjs
