+++ views/layouts/default.hbs
@@ -26,7 +26,7 @@
     </head>
     <body>
         <header>
-            <h1>Header</h1>
+            <h1>ToDo App</h1>
         </header>
         <div class='content'>
             {{{body}}}
